requirements.txt
PyQt6>=6.6.0
Pillow>=10.4.0
numpy>=1.26.0,<3
pytest>=7.0.0
